refactor: add firebase real and stub implementations
